chore(http): split params types into a dedicated http-h2 crate
This sets up configuring ClientParams from the proxy-api-resolve crate, i.e. so
that the proxy-api-resolve crate does not need to depend on the proxy-http
crate.
